split of the config editor into its own file and added some documentation
